BUG: MGH File writing bug fixed
In order to identify and more carefully ensure
that conversions are done consistently for Reading/Writing,
the code was converted from naked c-arrays to itk::Vector and
itk::Matrix so that the manipulation of different spaces can
look very much like the documentation describing the conversion
between spaces.

This conversion also assists with debugging and future maintenance
by allowing out of code validation to be done in Matlab, Numpy, or
similar matrix algebra tools.